feat: Add Palindrome Number solution
Add a new file `Palindrome_Number_9.java` which contains a solution to LeetCode problem 9, "Palindrome Number".

The file includes a `palindrome` method that checks if a number is a palindrome by reversing it. It also contains a commented-out, more optimized `isPalindrome` method from a LeetCode `Solution` class that avoids potential integer overflow by only reversing half of the number.

Signed-off-by: Anant Mittal <anantmittal943@gmail.com>