feat: Add Valid Palindrome solution
Add a new file `Valid_Palindrome_125.java` which contains a solution to LeetCode problem 125, "Valid Palindrome".

The file includes an `isPalindrome` method that checks if a given string is a palindrome by ignoring non-alphanumeric characters and case. The implementation uses a two-pointer approach to compare characters from both ends of the string. The file also contains a commented-out `Solution` class with an identical implementation.

Signed-off-by: Anant Mittal <anantmittal943@gmail.com>